+++ mongodb/src/main/scala/quasar/physical/mongodb/planner/exprOp.scala
@@ -30,7 +30,7 @@ import matryoshka.implicits._
 import matryoshka.patterns._
 import scalaz._, Scalaz._
 
-object exprHelpers {
+object exprOp {
   def processMapFuncExpr
     [T[_[_]]: BirecursiveT: ShowT, M[_]: Monad, EX[_]: Traverse, A]
     (funcHandler: AlgebraM[M, MapFunc[T, ?], Fix[EX]], staticHandler: StaticHandler[T, EX])
