+++ mongodb/src/main/scala/quasar/physical/mongodb/planner/javascript.scala
@@ -29,7 +29,7 @@ import matryoshka.implicits._
 import matryoshka.patterns._
 import scalaz._, Scalaz._
 
-object jsHelpers {
+object javascript {
   def processMapFunc[T[_[_]]: BirecursiveT: ShowT, M[_]: Monad: MonadFsErr: ExecTimeR, A]
     (fm: FreeMapA[T, A])(recovery: A => JsCore)
       : M[JsCore] =
